test(MenuGroup): 메뉴그룹 생성 RestAssured 기반 테스트로 변경
